pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = "SLAS - SLPortal Backend"
authors = ["Danny Rehl", "Dustin Heckmann"]
license = "GPLv3"
package-mode=false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {version = "^0.116", extras = ["all"]}
minio = "^7.2"
psycopg2-binary = "^2.9"
pydantic-settings = "^2.8"
pyjwt = {version = "^2.10", extras = ["crypto"]}
python-multipart = "^0.0"
requests = "^2.32"
slowapi = "^0.1"
sqlalchemy = "^2.0"

[tool.poetry.group.dev.dependencies]
alembic = "^1.15, !=1.15.0"
bandit = "^1.8"
coverage = "^7.7"
freezegun = "^1.5"
gevent = "^24.11"
isort = "^6.0"
mypy = "^1.15"
pylint = "^3.3"
pytest = "^8.3"
pytest-asyncio = "^0.26"
pytz = "^2025.2"
testcontainers = "^4.9"

[tool.bandit]
exclude_dirs = ["tests", ".venv"]

[tool.coverage.run]
concurrency = ["thread", "gevent"]
omit = [
    "tests/*",
    "*/tests/*"
]

[tool.coverage.report]
omit = ["**/tests/*"]

[tool.isort]
line_length = 100
multi_line_output=5

[[tool.mypy.overrides]]
module = "fastapi_auth_middleware"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "minio"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "minio.datatypes"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "minio.error"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "decouple"
ignore_missing_imports = true

[tool.pylint.basic]
good-names= """
    id,
    _,
"""

[tool.pylint."MESSAGES CONTROL"]
disable= """
    missing-class-docstring,
    missing-function-docstring,
    missing-module-docstring,
    too-few-public-methods,
    too-many-instance-attributes,
    too-many-arguments,
    too-many-positional-arguments,
    duplicate-code,
    fixme,
"""

[tool.pylint.format]
max-line-length = 120

[tool.pylint.master]
extension-pkg-whitelist = "pydantic"

[tool.pytest]
collect_ignore_glob = [
    "**/multimedia_files.py::NotAsJson",
    "**/db_migrations/env.py"
]

[tool.pytest.ini_options]
addopts = [
    "--import-mode=importlib",
    "-rA",
    "--tb=long",
    "--junitxml=./pytest-results.xml"
]
console_output_style = "count"
junit_family = "xunit2"
asyncio_default_fixture_loop_scope = "function"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
